requirements.txt
mal-api
pyinstaller
